Example of in-band registration of user by admin
XEP-0077 provides a specification for the in-band creation of new users, and this is usually done by an unauthenticated client (since the user who will be authenticating does not yet exist). It may be desirable in a client application to allow *only* authenticated administrators (or bots with administrative privileges) to create new accounts on behalf of other users.

This example, when paired with the appropriate modifications on the server, demonstrates how this works.

I probably wouldn't have gotten this working without legastero's help here:
	http://groups.google.com/group/sleekxmpp-discussion/browse_thread/thread/f2652563db0def11

The corresponding server changes for ejabberd are here:
	https://github.com/lehrblogger/ejabberd/commit/d58f21a404aef6394a193ee8c3320a08eb2f6e1b

The following links may also be useful context:
	http://www.ejabberd.im/node/5020
	http://www.ejabberd.im/node/1211
	http://yo.jabber.ru/bugzilla/show_bug.cgi?id=300
	http://stackoverflow.com/questions/7709658/asmack-xmpp-new-user-registration/7725400